test: rename form-input test suite to validation
